Friendly URL - Events no longer need a /details. Also made browse it's own controller
