+++ Epilogger.Web/Views/Browse/_BrowseEventOverview.cshtml
@@ -70,7 +70,7 @@
                     }
                 </div>
                 <div class="hottest-Title">
-                    <a href="/Events/@item.Event.EventSlug/details">
+                    <a href="/Events/@item.Event.EventSlug">
                         <h1>@item.Event.Name</h1>
                     </a>
                     <div class="hottest-totals">
